INF-1254
Add a managed-branch function that determines whether a Stash branch
should be managed by Stashkins.  A branch is managed if its name
starts with "feature/".  This gates whether a Jenkins jobs is deleted
if its branch is missing, and whether a job is created if the Stash
branch starts with "feature/".